+++ caqtus/gui/__init__.py
@@ -3,6 +3,6 @@
 It is implemented as a separate package build on top of the core functionalities.
 """
 
-from . import condetrol, snapshot
+from . import condetrol, snapshot, autogen
 
-__all__ = ["condetrol", "snapshot"]
+__all__ = ["condetrol", "snapshot", "autogen"]