+++ tests/test_gui/test_condetrol/test_device_configuration_editors/test_autogen/test_camera_editor.py
@@ -4,9 +4,7 @@ import attrs
 from pytestqt.qtbot import QtBot
 
 from caqtus.device.camera import CameraConfiguration
-from caqtus.gui.condetrol.device_configuration_editors._autogen import (
-    build_device_configuration_editor,
-)
+from caqtus.gui.autogen import build_device_configuration_editor
 from caqtus.types.image import Width, Height
 from caqtus.types.image.roi import RectangularROI